# `MyPy` and `Pylint` partition inputs via `CoarsenedTarget` (Cherry-pick of #15141)